refactor: code shuffle success page view
