+++ src/components/ProjectDashboard/components/SuccessPayView/index.ts
@@ -1,0 +1,1 @@
+export * from './SuccessPayView'
